+++ packages/framework/src/Model/Payment/Grid/PaymentGridFactory.php
@@ -7,8 +7,8 @@ use Shopsys\FrameworkBundle\Component\Grid\GridFactory;
 use Shopsys\FrameworkBundle\Component\Grid\GridFactoryInterface;
 use Shopsys\FrameworkBundle\Component\Grid\QueryBuilderWithRowManipulatorDataSource;
 use Shopsys\FrameworkBundle\Model\Localization\Localization;
-use Shopsys\FrameworkBundle\Model\Payment\Detail\PaymentDetailFactory;
 use Shopsys\FrameworkBundle\Model\Payment\Payment;
+use Shopsys\FrameworkBundle\Model\Payment\PaymentPriceCalculation;
 use Shopsys\FrameworkBundle\Model\Payment\PaymentRepository;
 
 class PaymentGridFactory implements GridFactoryInterface
@@ -26,25 +26,25 @@ class PaymentGridFactory implements GridFactoryInterface
     private $paymentRepository;
 
     /**
-     * @var \Shopsys\FrameworkBundle\Model\Payment\Detail\PaymentDetailFactory
-     */
-    private $paymentDetailFactory;
-
-    /**
      * @var \Shopsys\FrameworkBundle\Model\Localization\Localization
      */
     private $localization;
 
+    /**
+     * @var \Shopsys\FrameworkBundle\Model\Payment\PaymentPriceCalculation
+     */
+    private $paymentPriceCalculation;
+
     public function __construct(
         GridFactory $gridFactory,
         PaymentRepository $paymentRepository,
-        PaymentDetailFactory $paymentDetailFactory,
-        Localization $localization
+        Localization $localization,
+        PaymentPriceCalculation $paymentPriceCalculation
     ) {
         $this->gridFactory = $gridFactory;
         $this->paymentRepository = $paymentRepository;
-        $this->paymentDetailFactory = $paymentDetailFactory;
         $this->localization = $localization;
+        $this->paymentPriceCalculation = $paymentPriceCalculation;
     }
 
     /**
@@ -61,7 +61,7 @@ class PaymentGridFactory implements GridFactoryInterface
             'p.id',
             function ($row) {
                 $payment = $this->paymentRepository->findById($row['p']['id']);
-                $row['paymentDetail'] = $this->paymentDetailFactory->createDetailForPayment($payment);
+                $row['displayPrice'] = $this->getDisplayPrice($payment);
                 return $row;
             }
         );
@@ -70,7 +70,7 @@ class PaymentGridFactory implements GridFactoryInterface
         $grid->enableDragAndDrop(Payment::class);
 
         $grid->addColumn('name', 'pt.name', t('Name'));
-        $grid->addColumn('price', 'paymentDetail', t('Price'));
+        $grid->addColumn('price', 'displayPrice', t('Price'));
 
         $grid->setActionColumnClassAttribute('table-col table-col-10');
         $grid->addEditActionColumn('admin_payment_edit', ['id' => 'p.id']);
@@ -83,5 +83,15 @@ class PaymentGridFactory implements GridFactoryInterface
         );
 
         return $grid;
+    }
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Model\Payment\Payment $payment
+     * @return string
+     */
+    private function getDisplayPrice(Payment $payment)
+    {
+        $paymentBasePricesIndexedByCurrencyId = $payment->getBasePricesIndexedByCurrencyId($this->paymentPriceCalculation);
+        return $paymentBasePricesIndexedByCurrencyId[self::CURRENCY_ID_FOR_LIST]->getPriceWithVat();
     }
 }